+++ src/eval_f64/parser.rs
@@ -1,4 +1,5 @@
 use std::fmt;
+use std::sync::Arc;
 
 use super::ast::Node;
 use super::token::{NativeFunction, OperPrec, Token};
@@ -206,7 +207,7 @@ impl<'a> Parser<'a> {
                                 "There's no arguments in the min function".to_string(),
                             ));
                         }
-                        Node::Min(args)
+                        Node::Min(Arc::new(args))
                     }
                     NativeFunction::Max => {
                         let args = self.function_arguments()?;
@@ -215,7 +216,15 @@ impl<'a> Parser<'a> {
                                 "There's no arguments in the max function".to_string(),
                             ));
                         }
-                        Node::Max(args)
+                        Node::Max(Arc::new(args))
+                    }
+                    NativeFunction::Avg => {
+                        let args = self.function_arguments()?;
+                        if args.is_empty() {
+                            Node::Number(0.0)
+                        } else {
+                            Node::Avg(Arc::new(args))
+                        }
                     }
                 };
                 self.implicit_multiply(current_function)
@@ -246,11 +255,6 @@ impl<'a> Parser<'a> {
                 OperPrec::DefaultZero,
                 Token::RightParen,
                 |expr| expr,
-            ),
-            Token::Bar => self.get_enclosed_elements_with_impl_mult(
-                OperPrec::DefaultZero,
-                Token::Bar,
-                |expr| Node::Abs(Box::new(expr)),
             ),
             Token::LeftFloor => self.get_enclosed_elements_with_impl_mult(
                 OperPrec::DefaultZero,
@@ -482,12 +486,6 @@ mod tests {
         assert_eq!(parser.parse().unwrap(), expected);
     }
     #[test]
-    fn test_bar() {
-        let mut parser = Parser::new("|5|", None).unwrap();
-        let expected = Abs(Box::new(Number(5.0)));
-        assert_eq!(parser.parse().unwrap(), expected);
-    }
-    #[test]
     fn test_floor() {
         let mut parser = Parser::new("⌊5.25⌋", None).unwrap();
         let expected = Floor(Box::new(Number(5.25)));
@@ -658,25 +656,37 @@ mod tests {
     #[test]
     fn test_min_function() {
         let mut parser = Parser::new("min(3)", None).unwrap();
-        let expected = Min(vec![Number(3.0)]);
+        let expected = Min(Arc::new(vec![Number(3.0)]));
         assert_eq!(parser.parse().unwrap(), expected);
     }
     #[test]
     fn test_min_function2() {
         let mut parser = Parser::new("min(2,3,5)", None).unwrap();
-        let expected = Min(vec![Number(2.0), Number(3.0), Number(5.0)]);
+        let expected = Min(Arc::new(vec![Number(2.0), Number(3.0), Number(5.0)]));
         assert_eq!(parser.parse().unwrap(), expected);
     }
     #[test]
     fn test_max_function() {
         let mut parser = Parser::new("max(3)", None).unwrap();
-        let expected = Max(vec![Number(3.0)]);
+        let expected = Max(Arc::new(vec![Number(3.0)]));
         assert_eq!(parser.parse().unwrap(), expected);
     }
     #[test]
     fn test_max_function2() {
         let mut parser = Parser::new("max(2,3,5)", None).unwrap();
-        let expected = Max(vec![Number(2.0), Number(3.0), Number(5.0)]);
+        let expected = Max(Arc::new(vec![Number(2.0), Number(3.0), Number(5.0)]));
+        assert_eq!(parser.parse().unwrap(), expected);
+    }
+    #[test]
+    fn test_avg_function() {
+        let mut parser = Parser::new("avg(3)", None).unwrap();
+        let expected = Avg(Arc::new(vec![Number(3.0)]));
+        assert_eq!(parser.parse().unwrap(), expected);
+    }
+    #[test]
+    fn test_avg_function2() {
+        let mut parser = Parser::new("avg(2,3,5)", None).unwrap();
+        let expected = Avg(Arc::new(vec![Number(2.0), Number(3.0), Number(5.0)]));
         assert_eq!(parser.parse().unwrap(), expected);
     }
     #[test]
@@ -749,24 +759,6 @@ mod tests {
             Box::new(Ceil(Box::new(Number(2.0)))),
             Box::new(Floor(Box::new(Number(3.0)))),
         );
-        assert_eq!(parser.parse().unwrap(), expected);
-    }
-    #[test]
-    fn test_implicit_mul_abs_prts() {
-        let mut parser = Parser::new("|2|(3)", None).unwrap();
-        let expected = Multiply(Box::new(Abs(Box::new(Number(2.0)))), Box::new(Number(3.0)));
-        assert_eq!(parser.parse().unwrap(), expected);
-    }
-    #[test]
-    fn test_no_implicit_mul_abs() {
-        let mut parser = Parser::new("|2||3|", None).unwrap();
-        let expected = Abs(Box::new(Number(2.0)));
-        assert_eq!(parser.parse().unwrap(), expected);
-    }
-    #[test]
-    fn test_no_implicit_mul_prts_abs() {
-        let mut parser = Parser::new("(2)|3|", None).unwrap();
-        let expected = Number(2.0);
         assert_eq!(parser.parse().unwrap(), expected);
     }
 }