+++ src/eval_i64/ast.rs
@@ -1,4 +1,4 @@
-use std::error;
+use std::{error, sync::Arc};
 
 #[derive(Debug, Clone, PartialEq)]
 pub enum Node {
@@ -21,8 +21,9 @@ pub enum Node {
     Exp(Box<Node>),
     Exp2(Box<Node>),
     Sign(Box<Node>),
-    Min(Vec<Node>),
-    Max(Vec<Node>),
+    Min(Arc<Vec<Node>>),
+    Max(Arc<Vec<Node>>),
+    Avg(Arc<Vec<Node>>),
     Root(Box<Node>, Box<Node>),
     Log(Box<Node>, Box<Node>),
     Number(i64),
@@ -93,9 +94,9 @@ pub fn eval(expr: Node) -> Result<i64, Box<dyn error::Error>> {
         }
         Min(args) => {
             if args.len() > 1 {
-                let mut result = i64::MAX;
-                for arg in args {
-                    result = eval(arg)?.min(result);
+                let mut result = i64::MIN;
+                for arg in <Vec<Node> as Clone>::clone(&args).into_iter() {
+                    result = eval(arg).unwrap().min(result);
                 }
                 Ok(result)
             } else {
@@ -107,9 +108,9 @@ pub fn eval(expr: Node) -> Result<i64, Box<dyn error::Error>> {
         }
         Max(args) => {
             if args.len() > 1 {
-                let mut result = i64::MIN;
-                for arg in args {
-                    result = eval(arg)?.max(result);
+                let mut result = i64::MAX;
+                for arg in <Vec<Node> as Clone>::clone(&args).into_iter() {
+                    result = eval(arg).unwrap().max(result);
                 }
                 Ok(result)
             } else {
@@ -118,6 +119,13 @@ pub fn eval(expr: Node) -> Result<i64, Box<dyn error::Error>> {
                     None => Ok(0),
                 }
             }
+        }
+        Avg(args) => {
+            let mut result = 0;
+            for arg in <Vec<Node> as Clone>::clone(&args).into_iter() {
+                result += eval(arg).unwrap();
+            }
+            Ok(result / (args.len() as i64))
         }
     }
 }