+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2024-06-30
+
+Removal of Bar - the absolute value symbol - (too unstable).
+Optimize vector use with `Arc`.
+Add `AVG` function to `eval_f64` and `eval_i64`.
+
 ## [0.1.1] - 2024-06-30
 
 Some minor fixes and tests has been introduced.